requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiomysql>=0.2.0
PyJWT>=2.8.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
